+++ src/literalizer/languages/cobol.py
@@ -36,6 +36,7 @@ from literalizer._formatters.format_integers import (
 )
 from literalizer._language import (
     CallStyle,
+    CallSupport,
     CommentConfig,
     DateFormatConfig,
     DatetimeFormatConfig,
@@ -519,7 +520,9 @@ class Cobol(metaclass=LanguageCls):
     static_preamble: ClassVar[Sequence[str]] = ()
     static_body_preamble: ClassVar[Sequence[str]] = ()
     special_float_preamble: ClassVar[tuple[str, ...]] = ()
-    call_style_config: ClassVar[CallStyle | None] = None
+    call_style_config: ClassVar[CallStyle | CallSupport] = (
+        CallSupport.NOT_IMPLEMENTED_BY_TOOL
+    )
 
     @cached_property
     def format_string(self) -> Callable[[str], str]: